raspi_client/requirements.txt
# 라즈베리파이 클라이언트 필수 패키지
requests>=2.28.0

# 이미지 처리 (줌인 효과용)
Pillow>=9.0.0

# TTS (음성 출력)
gTTS>=2.3.0

# 라즈베리파이 카메라 (라즈베리파이 OS에 기본 포함)
# picamera2

# 참고: 라즈베리파이에서 추가로 설치해야 할 것들
# sudo apt-get install mpg321  (또는 mpg123)
# sudo apt-get install ffmpeg  (aplay 사용시)
